Conflicts:
	.gitignore
	catroid/src/org/catrobat/catroid/content/bricks/ChangeBrightnessByNBrick.java
	catroid/src/org/catrobat/catroid/content/bricks/ChangeGhostEffectByNBrick.java
	catroid/src/org/catrobat/catroid/content/bricks/SetBrightnessBrick.java
	catroid/src/org/catrobat/catroid/content/bricks/SetGhostEffectBrick.java
	catroid/src/org/catrobat/catroid/livewallpaper/WallpaperCostume.java

+++ catroid/src/org/catrobat/catroid/content/bricks/ChangeGhostEffectByNBrick.java
@@ -135,7 +135,7 @@ public class ChangeGhostEffectByNBrick implements Brick, OnClickListener {
 		}
 
 		float changePercentage = (float) changeGhostEffect / 100;
-		int alpha = wallpaperCostume.getAlphaValue() - (int) (changePercentage * 255);
+		int alpha = (int) (wallpaperCostume.getAlphaValue() - (int) (changePercentage * 255));
 
 		wallpaperCostume.setAlphaValue(alpha);
